api_wrapper/requirements.txt
cython
falcon
python-powerdns
requests